feature: Fully incorporate lexer into parser
Replace ranges in AST nodes with `AstInfo`,
that keeps the consumed tokens.
Consume comment tokens in front of each other token.
This way, they are also in the `AstInfo`
and can be used in hover help etc.
Fix offset logic for `Tokens`.
Support slicing for `Tokens` and use it for AstInfo.
Add `ToRange` derive macro and use it for AST nodes.
Adjust the usage of crate `spl_frontend` in lsp4spl.
Replace `ident_at` with a functions that finds the ident in tokens
instead of AST.
All tests pass, except for `acker`,
where token slices still need to be fixed.

Resolves: #4 & #1