fix: enable cursor after last character
Apparently, VSCode is not POSIX-compliant,
so a line is not always terminated with a newline character.
This way, the (insertion) cursor can be
after the last character of a file
and therefore at a non-valid index of the text.
This fix allows this behaviour,
so that completion is possible
even if the cursor is behind the last character.

To reflect this behaviour,
the `get_index` function was renamed to `get_insertion_index`.