feature: configurable server executable
The VSCode extension can be run in debug and run mode.
In debug mode, it uses the debug binary from the project,
it lives in.
In run mode, the executable can be configured by a setting.
Added contribution point for this setting.
Show error message, if no path is specified.
Add vsce as a dependency, so that the extension can be packaged
and published.