refactor: reference adds offset to range
`Reference<T>`'s implementation of the trait `ToRange`
changes from just returning the range of it's contained type
to the range of the contained type plus the offset.
This makes it possible to use `to_range()` in generic contexts,
where the field `offset` is not available.

Parts of table building and semantic analysis
built up on the old implementation,
so they access the contained type now directly.