feature: add incremental parsing for global decs
Add incremental parsing for global declarations.

Change `Parser` trait.
It now takes an `Option<Self>` as input.
Add helper function `inc` to avoid closures with the new `parse` method.
Add own parser for `Reference<T>` as alternative to `reference`.

Add `TokenChange` as part of the token stream.
The `lexer::update` function calculates the token change.

Adjust `AnalyzedSource::update` to incremental parsing.

Add `affected` parser.
It checks whether the node was affected by the change.
If yes, it is tried to reanalyse it.
Except, if it was completely deleted, then a reanalysis is not useful.
If it was unaffected,
it can still be discarded,
if it's tokens were already partially consumed (indirectly affected).

Introduce own `ParserError`,
to be able to use own errors during parsing.
For example the `Affected` error.
`nom` errors are just wrapped.

Add own `many` parser.
It incrementally re-parses a vector of nodes.
It accounts for insertions
and does not fail on `Affected` errors.

Copy `nom::branch::alt` and adjust it so that it fails immediately
on `Affected` errors.
TODO: This might actually not be necessary.

Add tests that compare the incremental output,
to an from-scratch-output.
Add `insta` as dependency,
to control AST output of tests.