fix: prevent re-parsing by the wrong parser
Before, it was possible for a parser,
whose tokens were already partially consumed,
to parse the tokens of the next node,
if it was of the same kind.
Now all parsers whose tokens are already partially consumed
are marked as affected.

Add `TokenChange::new_token_pos` to translate
between old and new token positions.

Simplify condition for insertion parsing.