fix: remove old error messages from reused nodes
Build and semantic error messages should not persist in reused nodes.
They will be again be added in the respective phases,
if they are still valid.

Add trait `AstInfoTraverser` to remove the message from every sub node.

Note: Used a traversal with function pointers instead of Fn traits
to avoid a recursion problem in the compiler.
It might be possible to do this with normal traits as well.