fix: incremental expect
The `expect` parser did not try to reanalyze the input on an `AffectedError`.
This led to problems, when the expression was deleted but replaced.
`expect` is now adapted and takes a `this` parameter, too.

`inc` parsers meaning was changed, to add closure with `this` parameter.
Tests with original `inc` parser were modified to use other closure.